changed entity location to 'domain'
