+++ src/main/java/org/project/skyflow/domain/entity/Content.java
@@ -1,8 +1,8 @@
-package org.project.skyflow.entity;
+package org.project.skyflow.domain.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
-import org.project.skyflow.entity.type.ContentStatus;
+import org.project.skyflow.domain.entity.type.ContentStatus;
 
 import java.time.LocalDateTime;
 import java.util.List;
