+++ src/main/java/org/project/skyflow/domain/entity/type/ContentStatus.java
@@ -1,4 +1,4 @@
-package org.project.skyflow.entity.type;
+package org.project.skyflow.domain.entity.type;
 
 public enum ContentStatus {
     PUBLIC,
