+++ src/main/java/org/project/skyflow/domain/entity/type/FollowType.java
@@ -1,0 +1,6 @@
+package org.project.skyflow.domain.entity.type;
+
+public enum FollowType {
+    ALL,
+    NONE
+}
