+++ src/main/java/org/project/skyflow/domain/entity/type/ProcessingStatus.java
@@ -1,4 +1,4 @@
-package org.project.skyflow.entity.type;
+package org.project.skyflow.domain.entity.type;
 
 public enum ProcessingStatus {
     UPLOADED,
