+++ src/main/java/org/project/skyflow/domain/entity/type/Role.java
@@ -1,0 +1,6 @@
+package org.project.skyflow.domain.entity.type;
+
+public enum Role {
+    USER,
+    CREATOR
+}
