--- src/main/java/org/project/skyflow/entity/type/FollowType.java
@@ -1,6 +1,0 @@
-package org.project.skyflow.entity.type;
-
-public enum FollowType {
-    ALL,
-    NONE
-}
