+++ src/main/java/org/project/skyflow/repository/ContentMetaDataRepository.java
@@ -1,6 +1,6 @@
 package org.project.skyflow.repository;
 
-import org.project.skyflow.entity.ContentMetaData;
+import org.project.skyflow.domain.entity.ContentMetaData;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
