+++ src/main/java/org/project/skyflow/repository/NotificationRepository.java
@@ -1,6 +1,6 @@
 package org.project.skyflow.repository;
 
-import org.project.skyflow.entity.Notification;
+import org.project.skyflow.domain.entity.Notification;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
