+++ src/main/java/org/project/skyflow/repository/OpinionRepository.java
@@ -1,6 +1,6 @@
 package org.project.skyflow.repository;
 
-import org.project.skyflow.entity.Opinion;
+import org.project.skyflow.domain.entity.Opinion;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
