+++ src/main/java/org/project/skyflow/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package org.project.skyflow.repository;
 
-import org.project.skyflow.entity.User;
+import org.project.skyflow.domain.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
